fix: Stop the scripts if upload and download fail
Re-raise any exceptions that happen when interacting with S3 to stop
the script at the correct place and provide better stack traces.
Without this re-raise, in case of the restore, the script would
continue even if the file integrity check failed and would break on
extracting the non-existing tar file.