feat: Add ability to enable and disable CronJobs
Previously we relied on the existence of the schedule configuration
parameters to add or remove the corresponding backup and restore
CronJob. But, if a CronJob is already defined, removing its manifest
does not deactivate it. To deactivate a CronJob we need to suspend it.
Add `BACKUP_K8S_CRONJOB_BACKUP_ENABLE` and
`BACKUP_K8S_CRONJOB_RESTORE_ENABLE` to set the `.spec.suspend` in the
CronJob manifest. We also need to set `.spec.startingDeadlineSeconds`
to prevent the missed jobs to be immediately scheduled when
`.spec.suspend` is toggled from `true` to `false`.

Reference:
https://kubernetes.io/docs/tasks/job/automated-tasks-with-cron-jobs/