fix: Use ephemeral volumes in K8s
The backup plugin uses the `data` directory to store the db dumps
and the compressed backup file.
In a local deployment, it uses the local storage.
However, in K8s, using the node's local is not a good practice.
Instead, we configure an ephemeral volume that is created and destroyed
together with the job.