feat: Add ability to select MongoDB databases
Add the `BACKUP_MONGODB_DATABASES` setting to select the
MongoDB databases to back up. If this setting is not provided,
back up all databases.

Since we are providing the `--db` flag to `mongodump`, when backing
up specific databases, we need to also provide the authentication
database. Otherwise `mongodump` uses whatever database is set by `--db`
to authenticate. To give users the ability to specify the
authentication database, add `BACKUP_MONGODB_AUTHENTICATION_DATABASE`
which defaults to `admin`.