fix: Flush logs after MySQL restore
If we are restoring a large database, the large amount of statements
executed my importing a mysqldump may lead to a considerable number of
binlog entries. Afterwards, it is usually prudent to have MySQL remove
expired binlog files. One way to force MySQL into doing that is to
flush logs.

Thus, add a "mysqladmin flush-logs" command right after successfully
completing a restore.